// src/app/Interfaces/user-profile.ts
export interface UserProfile {
    userName: string;
    email: string;
    phoneNumber: string | null;
    secondPhoneNumber: string | null;
    age: number | null;
    gender: number | null;
    maritalStatus: number | null;
    education: number | null;
    employment: string | null;
    job: string | null;
}